Partially fix PUT and POST requests in BadMethods unit tests
Dummy JSON data added to PUT and POST requests in the unit tests to work
around a misdetection of the MIME type by piston.

All piston resources marked csrf_exempt to prevent django's CsrfMiddleware from
interfering with them.

For 3 of the 4 BadMethods tests still fail because with an OAuth error when
using POST.