Initial HTML5 offline note app built on jQuery Mobile
After logging in, hit <domain>/mobile/ in a webkit-based browser.

Major missing features are transforming note content to HTML,
automatic synchronization, and, of course, editing.

One major thing to change is that we are bundling a recent git
version of jquery mobile and related assets. This is just my own
convenience, for now.