[mobile] Only load 10 notes at a time, use button to load more
May change this in the future to auto-load 10 more notes when
scroll hits bottom of list.  If so, I'd remove the button and
turn of data-inset on the on the note list.

This is done for performance reasons.  Even on my Core i7 laptop,
it took 4 seconds for index.html to finish loading, all because
of the processing time jQuery spends putting my ~600 notes in the
note list.